refactor GPTQ: move all GPTQ files to GPTQ folder
