+++ model_compression_toolkit/gptq/keras/gptq_training.py
@@ -24,16 +24,16 @@ else:
     from keras.engine.base_layer import TensorFlowOpLayer
 
 from model_compression_toolkit.core import common
-from model_compression_toolkit.core.common.gptq.gptq_training import GPTQTrainer
-from model_compression_toolkit.core.common.gptq.gptq_config import GradientPTQConfig
+from model_compression_toolkit.gptq.common.gptq_training import GPTQTrainer
+from model_compression_toolkit.gptq.common.gptq_config import GradientPTQConfig
 from model_compression_toolkit.core.common import Graph
-from model_compression_toolkit.core.keras.gradient_ptq.graph_info import get_trainable_parameters, get_weights_for_loss
+from model_compression_toolkit.gptq.keras.graph_info import get_trainable_parameters, get_weights_for_loss
 from model_compression_toolkit.core.common.framework_info import FrameworkInfo
 from model_compression_toolkit.core.common.framework_implementation import FrameworkImplementation
 import numpy as np
 import copy
 from model_compression_toolkit.core.keras.constants import BIAS, USE_BIAS
-from model_compression_toolkit.core.keras.quantizer.gradient_ptq import WeightQuantizeConfig
+from model_compression_toolkit.gptq.keras.quantizer import WeightQuantizeConfig
 
 
 class KerasGPTQTrainer(GPTQTrainer):
