+++ model_compression_toolkit/gptq/keras/quantization_facade.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 # ==============================================================================
 
-from typing import Callable, List
+from typing import Callable
 
 from model_compression_toolkit.core import common
 from model_compression_toolkit.core.common import Logger
 from model_compression_toolkit.core.common.constants import TENSORFLOW
-from model_compression_toolkit.core.common.gptq.gptq_config import GradientPTQConfig
+from model_compression_toolkit.gptq.common.gptq_config import GradientPTQConfig
 from model_compression_toolkit.core.common.mixed_precision.kpi import KPI
 from model_compression_toolkit.core.common.framework_info import FrameworkInfo
 from model_compression_toolkit.core.common.mixed_precision.mixed_precision_quantization_config import \
@@ -38,7 +38,7 @@ if importlib.util.find_spec("tensorflow") is not None\
     from model_compression_toolkit.core.keras.keras_implementation import KerasImplementation
     from model_compression_toolkit.core.keras.keras_model_validation import KerasModelValidation
     from tensorflow.keras.models import Model
-    from model_compression_toolkit.core.keras.gradient_ptq.gptq_loss import multiple_tensors_mse_loss
+    from model_compression_toolkit.gptq.keras.gptq_loss import multiple_tensors_mse_loss
     from keras.optimizer_v2.optimizer_v2 import OptimizerV2
     from model_compression_toolkit.core.keras.constants import DEFAULT_TP_MODEL
