+++ model_compression_toolkit/gptq/keras/quantizer/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Sony Semiconductors Israel, Inc. All rights reserved.
+# Copyright 2022 Sony Semiconductors Israel, Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,3 +13,4 @@
 # limitations under the License.
 # ==============================================================================
 
+from model_compression_toolkit.gptq.keras.quantizer.weight_quantizer_gptq_config import WeightQuantizeConfig
