+++ model_compression_toolkit/gptq/keras/quantizer/utils.py
